Refactor SelectedVideo provider
Initial use state is passed inside the provider file
