+++ src/providers/SelectedVideo/index.js
@@ -1,0 +1,1 @@
+export { default } from './SelectedVideo.provider';
